refactor(service): InputValidationService 인터페이스 적용 - 서비스의 유연성을 위해서 인터페이스 적용 - 의미를 보다 명확하게 만들기 위해 이름 변경
